+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### [0.1.10] - 2026-08-14
 - Add filters to statistics view.
-- 
+- Added referral codes to users.
 
 ### [0.1.9] - 2026-08-12
 - Adjusted code generation in order creation to include a prefix based on the client type.